+++ src/main/kotlin/com/github/victorrentea/livecoding/lombok/LombokAndHibernate.kt
@@ -6,9 +6,13 @@ import com.intellij.codeInspection.LocalInspectionTool
 import com.intellij.codeInspection.LocalQuickFixOnPsiElement
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
+import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.command.WriteCommandAction
 import com.intellij.openapi.project.Project
 import com.intellij.psi.*
+import com.intellij.psi.codeStyle.CodeStyleManager
 import com.intellij.psi.codeStyle.JavaCodeStyleManager
+import com.intellij.refactoring.extractMethod.newImpl.ExtractMethodHelper.addSiblingAfter
 import org.jetbrains.annotations.NotNull
 
 class LombokAndHibernateInspection : LocalInspectionTool() {
@@ -40,8 +44,8 @@ class LombokAndHibernateVisitor(private val holder: ProblemsHolder) : PsiElement
             // check for children
             if (psiClass.fields.any {
                     it.hasCollectionType() &&
-                    !it.hasAnnotation("lombok.ToString.Exclude") &&
-                    !it.hasAnnotation("lombok.ToString.Include")
+                            !it.hasAnnotation("lombok.ToString.Exclude") &&
+                            !it.hasAnnotation("lombok.ToString.Include")
                 }) {
                 holder.registerProblem(
                     it,
@@ -60,20 +64,24 @@ class LombokAndHibernateVisitor(private val holder: ProblemsHolder) : PsiElement
                 RemoveHashCodeEqualsQuickFix(psiClass)
             )
         }
-
-
     }
-
 }
 
 /** Adds the annotation if not already present */
-fun PsiModifierListOwner.setAnnotation(qualifiedName:String) {
-    val modifiers = modifierList ?: return
-    if (modifiers.hasAnnotation(qualifiedName)) return
+fun PsiModifierListOwner.setAnnotation(
+    qualifiedName: String,
+    shortenFQName: Boolean = true
+): PsiAnnotation? {
+    val modifiers = modifierList ?: return null
+    if (modifiers.hasAnnotation(qualifiedName)) return null
     val annotation = modifiers.addAnnotation(qualifiedName)
-    JavaCodeStyleManager.getInstance(project).shortenClassReferences(annotation)
+    if (shortenFQName) {
+        JavaCodeStyleManager.getInstance(project).shortenClassReferences(annotation)
+    }
+    return annotation
 }
-fun PsiModifierListOwner.removeAnnotation(qualifiedName:String) {
+
+fun PsiModifierListOwner.removeAnnotation(qualifiedName: String) {
     val modifiers = modifierList ?: return
     val annotation = modifiers.findAnnotation(qualifiedName) ?: return
     annotation.delete()
@@ -89,12 +97,21 @@ class ReplaceDataOnEntityQuickFix(psiClass: PsiClass) : LocalQuickFixOnPsiElemen
     override fun getText() = "Replace @Data with safer parts"
 
     override fun invoke(project: Project, file: PsiFile, psiClass: PsiElement, endElement: PsiElement) {
-        if (psiClass !is PsiClass) return;
-        psiClass.removeAnnotation("lombok.Data")
-        psiClass.setAnnotation("lombok.Getter")
-        psiClass.setAnnotation("lombok.Setter")
-        psiClass.setAnnotation("lombok.ToString")
-        ExcludeCollectionsFromToStringQuickFix.excludeCollectionFields(psiClass)
+        if (psiClass !is PsiClass) return
+        ApplicationManager.getApplication().invokeLater {
+            WriteCommandAction.runWriteCommandAction(project, "Replace @Data With Safer Parts", "Live-Coding", {
+                val javaPsi = JavaPsiFacade.getInstance(project)
+                psiClass.removeAnnotation("lombok.Data")
+                psiClass.setAnnotation("lombok.Getter")
+                psiClass.setAnnotation("lombok.Setter")?.let {
+                    val commentOnSetter = javaPsi.elementFactory.createCommentFromText("// consider encapsulating changes", psiClass)
+                    it.addSiblingAfter(commentOnSetter)
+                }
+
+                psiClass.setAnnotation("lombok.ToString")
+                ExcludeCollectionsFromToStringQuickFix.excludeCollectionFields(psiClass)
+            })
+        }
     }
 }
 
@@ -112,7 +129,7 @@ class ExcludeCollectionsFromToStringQuickFix(psiClass: PsiClass) : LocalQuickFix
             // TODO what about fields from superclasses : Postpone as it's not likely to have a collection in supertype
             for (field in psiClass.fields) {
                 if (field.hasCollectionType() && !field.hasAnnotation("lombok.ToString.Include")) {
-                    field.setAnnotation("lombok.ToString.Exclude")
+                    field.setAnnotation("lombok.ToString.Exclude", false)
                 }
             }
         }