+++ src/test/live-coding-playground/src/main/java/lombokplay/addrac/AddRAC2.java
@@ -1,4 +1,4 @@
-package lombok.addrac;
+package lombokplay.addrac;
 
 public class AddRAC2 {
     private final int x; //