+++ src/test/live-coding-playground/src/main/java/lombokplay/replacerac/ReplaceRAC_TwoArg_Component.java
@@ -1,4 +1,4 @@
-package lombok.replacerac;
+package lombokplay.replacerac;
 
 import org.springframework.stereotype.Component;