+++ src/test/live-coding-playground/src/main/java/lombokplay/replacerac/ReplaceRAC_TwoArg_ExtraCode.java
@@ -1,4 +1,4 @@
-package lombok.replacerac;
+package lombokplay.replacerac;
 
 public class ReplaceRAC_TwoArg_ExtraCode {
     private final int dummy;