+++ src/test/live-coding-playground/src/main/java/lombokplay/replacerac/ReplaceRAC_notStraightAssign.java
@@ -1,4 +1,4 @@
-package lombok.replacerac;
+package lombokplay.replacerac;
 
 import java.util.Objects;