+++ src/test/live-coding-playground/src/main/java/lombokplay/slf4j/AddSlf4j.java
@@ -1,4 +1,4 @@
-package lombok.slf4j;
+package lombokplay.slf4j;
 
 
 import lombok.extern.slf4j.Slf4j;